requirements.txt
Flask==3.0.0
requests==2.31.0
google-generativeai==0.3.2
gunicorn==21.2.0

